Move AdditionalType class to internal
